Renamed main_common to args_common
since it only touches cli args related bits.
