+++ packages/ubuntu_wsl_setup/lib/main.dart
@@ -6,9 +6,9 @@ import 'package:subiquity_client/subiquity_server.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:ubuntu_wizard/app.dart';
 import 'package:ubuntu_wizard/utils.dart';
-import 'package:ubuntu_wsl_setup/main_common.dart';
 
 import 'app.dart';
+import 'args_common.dart';
 import 'installing_state.dart';
 
 Future<void> main(List<String> args) async {
